requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
scikit-image==0.21.0
tqdm==4.66.1